Fixed various GCC warnings
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@5762 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ networking/url.cpp
@@ -249,7 +249,7 @@ int URL::getFirstSlashPos(const std::string& urlstring)//returns -1 if no slash
 		}
 	}
 
-	if(first_slashpos == urlstring.length())
+	if(first_slashpos == (int)urlstring.length())
 	{
 		//no slash found
 		return -1;
